backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
reportlab==4.0.7

# Data processing
spacy==3.7.2
pandas==2.1.3
numpy==1.26.2

# Database
clickhouse-driver==0.2.6
clickhouse-connect

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utils
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
nox==2023.4.22
